Capturing Values
Converted numberical value to string in order to capture the length value.

Used template literals which allows you to add equations into strings, which converts the answer to the equation to a string value.

Example: `string ${a + b} string`
Be sure to use back-ticks instead of single or double quotes.